Fixed potential bug in limit value checking
The previous code checked that the value of `l` was in the range of `] -32 767; +32 768 [`.
This is wrong because on a 32-bit system such as the Teensy, the C standard defines that `int` is stored on 16 bits and `long` on 32 bits.
Therefore, the value of `l` must be in the range of `] -2^15; 2^15-1 [`, i.e `] -32 768; +32 767 [`, which wasn't the case.
In the edge case where `l` had the value of 32 768, this would return `i` as -32 768 (see two's complement representation of signed integers).
This bug was eliminated by using explicit-sized types defined in the `cstdint` header (`int16_t`, `uint32_t`, etc...) and their defined limit values (`INT16_MIN`, `UINT16_MAX`, etc...).